Factor out peephole optimizer Instead of removing instructions replace them with comments Run multiple passes of EliminatePushXXPopXX
